Adiciona serializers e viewasets
