+++ sme_portal_aluno_apps/alunos/models/log_consulta_eol.py
@@ -6,5 +6,5 @@ from sme_portal_aluno_apps.core.models_abstracts import ModeloBase
 
 class LogConsultaEOL(ModeloBase):
     codigo_eol = models.CharField(
-        "Código EOL do Aluno", max_length=7, unique=True, validators=[validators.MinLengthValidator(7)])
+        "Código EOL do Aluno", max_length=7, validators=[validators.MinLengthValidator(7)])
     json = JSONField('Log', blank=True, default=dict)
